# docs/openapi-baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gruff API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Gruff API Support"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "description": "Get a paginated list of users (admin only)",
        "operationId": "listUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsersResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/types": {
      "post": {
        "tags": [
          "Types"
        ],
        "summary": "Create type",
        "description": "Create a new entity or link type",
        "operationId": "createType",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateType"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Type created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Type"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Types"
        ],
        "summary": "List types",
        "description": "Get a paginated list of types",
        "operationId": "listTypes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category (entity or link)",
            "schema": {
              "type": "string",
              "enum": [
                "entity",
                "link"
              ]
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Filter by name (partial match)",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "List of types",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTypesResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/entities": {
      "post": {
        "tags": [
          "Entities"
        ],
        "summary": "Create entity",
        "description": "Create a new entity with the specified type and properties",
        "operationId": "createEntity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntity"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Entity created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Entity"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "List entities",
        "description": "Get a paginated list of entities with optional filters",
        "operationId": "listEntities",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type_id",
            "in": "query",
            "description": "Filter by entity type ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "created_by",
            "in": "query",
            "description": "Filter by creator user ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "created_after",
            "in": "query",
            "description": "Filter entities created after this timestamp (Unix ms)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "created_before",
            "in": "query",
            "description": "Filter entities created before this timestamp (Unix ms)",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/cursor"
          },
          {
            "$ref": "#/components/parameters/includeDeleted"
          }
        ],
        "responses": {
          "200": {
            "description": "List of entities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEntitiesResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/links": {
      "post": {
        "tags": [
          "Links"
        ],
        "summary": "Create link",
        "description": "Create a new link between two entities",
        "operationId": "createLink",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLink"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Link created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Link"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Links"
        ],
        "summary": "List links",
        "description": "Get a paginated list of links with optional filters",
        "operationId": "listLinks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type_id",
            "in": "query",
            "description": "Filter by link type ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "source_entity_id",
            "in": "query",
            "description": "Filter by source entity ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "target_entity_id",
            "in": "query",
            "description": "Filter by target entity ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/cursor"
          },
          {
            "$ref": "#/components/parameters/includeDeleted"
          }
        ],
        "responses": {
          "200": {
            "description": "List of links",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedLinksResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/search/entities": {
      "post": {
        "tags": [
          "Search"
        ],
        "summary": "Search entities",
        "description": "Search entities by JSON properties with advanced filtering",
        "operationId": "searchEntities",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchEntities"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEntitiesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/links": {
      "post": {
        "tags": [
          "Search"
        ],
        "summary": "Search links",
        "description": "Search links by JSON properties with advanced filtering",
        "operationId": "searchLinks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchLinks"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedLinksResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bulk/entities": {
      "post": {
        "tags": [
          "Bulk Operations"
        ],
        "summary": "Bulk create entities",
        "description": "Create multiple entities in a single request (max 100)",
        "operationId": "bulkCreateEntities",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateEntities"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk create results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkOperationResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Bulk Operations"
        ],
        "summary": "Bulk update entities",
        "description": "Update multiple entities in a single request (max 100)",
        "operationId": "bulkUpdateEntities",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateEntities"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk update results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkOperationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bulk/links": {
      "post": {
        "tags": [
          "Bulk Operations"
        ],
        "summary": "Bulk create links",
        "description": "Create multiple links in a single request (max 100)",
        "operationId": "bulkCreateLinks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateLinks"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk create results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkOperationResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Bulk Operations"
        ],
        "summary": "Bulk update links",
        "description": "Update multiple links in a single request (max 100)",
        "operationId": "bulkUpdateLinks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateLinks"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk update results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkOperationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /api/auth/login or /api/auth/register"
      }
    },
    "parameters": {
...
      "limit": {
        "name": "limit",
        "in": "query",
        "description": "Maximum items to return (1-100, default: 20)",
        "schema": {
          "type": "string",
          "default": "20"
        }
      },
      "cursor": {
        "name": "cursor",
        "in": "query",
        "description": "Pagination cursor from previous response",
        "schema": {
          "type": "string"
        }
      },
      "includeDeleted": {
        "name": "include_deleted",
        "in": "query",
        "description": "Include soft-deleted items (default: false)",
        "schema": {
          "type": "string",
          "default": "false"
        }
      }
    },
    "schemas": {
...
      "CreateType": {
        "type": "object",
        "required": [
          "name",
          "category"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Person"
          },
          "category": {
            "type": "string",
            "enum": [
              "entity",
              "link"
            ]
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          },
          "json_schema": {
            "type": "object",
            "description": "JSON Schema for property validation"
          }
        }
      },
...
      "PaginatedUsersResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "next_cursor": {
                "type": "string",
                "nullable": true
              },
              "total": {
                "type": "integer"
              }
            }
          }
        }
      },
...
    }
  }
}